Kernel: Move sys_read to correct file
